requirements.txt
streamlit==1.44.1
databricks-sdk
# databricks-sql-connector>=3.0.0  # Not available in runtime
mlflow-skinny>=2.21.2
pandas>=2.0.0
plotly>=5.0.0
pyspark>=3.3.0
delta-spark>=2.4.0
